refactor: test resources
- add adult features resource for testing
