+++ test/resources/adult/__init__.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+
+DIR = Path(__file__).parent
+
+
+PATH_STATS_CSV = DIR / "stats.csv"
