+++ test/resources/adult/stats.csv
@@ -1,0 +1,16 @@
+feature,missing_values,min,max,mean,std,1st_percentile,2nd_percentile,3rd_percentile,type,values,distribution,output,sensitive
+age,0,0.0,4.0,1.7695369510415284,1.266784781771372,1.0,2.0,3.0,integer,'[0, 1, 2, 3, 4]','{"keys": ["0.0 - 0.4", "0.4 - 0.8", "0.8 - 1.2", "1.2 - 1.6", "1.6 - 2.0", "2.0 - 2.4", "2.4 - 2.8", "2.8 - 3.2", "3.2 - 3.6", "3.6 - 4.0"], "values": [8441, 0, 12074, 0, 0, 11472, 0, 7936, 0, 5299]}',False,False
+workclass,0,,,,,,,,categorical,'["Federal-gov", "Local-gov", "Private", "Self-emp-inc", "Self-emp-not-inc", "State-gov", "Without-pay"]','{"keys": ["Private", "Self-emp-not-inc", "Local-gov", "State-gov", "Self-emp-inc", "Federal-gov", "Without-pay"], "values": [33307, 3796, 3100, 1946, 1646, 1406, 21]}',True,False
+fnlwgt,0,13492.0,1490400.0,189734.7343107337,105639.19513422118,117388.25,178316.0,237926.0,float,,'{"keys": ["13492.0 - 161182.8", "161182.8 - 308873.6", "308873.6 - 456564.4", "456564.4 - 604255.2", "604255.2 - 751946.0", "751946.0 - 899636.8", "899636.8 - 1047327.6", "1047327.6 - 1195018.4", "1195018.4 - 1342709.2", "1342709.2 - 1490400.0"], "values": [18701, 20916, 4803, 604, 140, 32, 12, 7, 3, 4]}',False,False
+education,0,,,,,,,,categorical,'["10th", "11th", "12th", "1st-4th", "5th-6th", "7th-8th", "9th", "Assoc-acdm", "Assoc-voc", "Bachelors", "Doctorate", "HS-grad", "Masters", "Preschool", "Prof-school", "Some-college"]','{"keys": ["HS-grad", "Some-college", "Bachelors", "Masters", "Assoc-voc", "11th", "Assoc-acdm", "10th", "7th-8th", "Others"], "values": [14783, 9899, 7570, 2514, 1959, 1619, 1507, 1223, 823, 3325]}',False,False
+education-num,0,1.0,16.0,10.118460041572686,2.5528811950758743,9.0,10.0,13.0,integer,'[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16]','{"keys": ["1.0 - 2.5", "2.5 - 4.0", "4.0 - 5.5", "5.5 - 7.0", "7.0 - 8.5", "8.5 - 10.0", "10.0 - 11.5", "11.5 - 13.0", "13.0 - 14.5", "14.5 - 16.0"], "values": [294, 449, 1499, 1223, 2196, 14783, 11858, 1507, 10084, 1329]}',False,False
+marital-status,0,,,,,,,,categorical,'["Divorced", "Married-AF-spouse", "Married-civ-spouse", "Married-spouse-absent", "Never-married", "Separated", "Widowed"]','{"keys": ["Married-civ-spouse", "Never-married", "Divorced", "Separated", "Widowed", "Married-spouse-absent", "Married-AF-spouse"], "values": [21055, 14598, 6297, 1411, 1277, 552, 32]}',False,False
+occupation,0,,,,,,,,categorical,'["Adm-clerical", "Armed-Forces", "Craft-repair", "Exec-managerial", "Farming-fishing", "Handlers-cleaners", "Machine-op-inspct", "Other-service", "Priv-house-serv", "Prof-specialty", "Protective-serv", "Sales", "Tech-support", "Transport-moving"]','{"keys": ["Craft-repair", "Prof-specialty", "Exec-managerial", "Adm-clerical", "Sales", "Other-service", "Machine-op-inspct", "Transport-moving", "Handlers-cleaners", "Others"], "values": [6020, 6008, 5984, 5540, 5408, 4808, 2970, 2316, 2046, 4122]}',False,False
+relationship,0,,,,,,,,categorical,'["Husband", "Not-in-family", "Other-relative", "Own-child", "Unmarried", "Wife"]','{"keys": ["Husband", "Not-in-family", "Own-child", "Unmarried", "Wife", "Other-relative"], "values": [18666, 11702, 6626, 4788, 2091, 1349]}',False,False
+race,0,,,,,,,,categorical,'["Amer-Indian-Eskimo", "Asian-Pac-Islander", "Black", "Other", "White"]','{"keys": ["White", "Black", "Asian-Pac-Islander", "Amer-Indian-Eskimo", "Other"], "values": [38903, 4228, 1303, 435, 353]}',False,True
+sex,0,,,,,,,,binary,'["Female", "Male"]','{"keys": ["Male", "Female"], "values": [30527, 14695]}',False,True
+capitalgain,0,0.0,4.0,0.20505506169563487,0.7561768730118918,0.0,0.0,0.0,integer,'[0, 1, 2, 3, 4]','{"keys": ["0.0 - 0.4", "0.4 - 0.8", "0.8 - 1.2", "1.2 - 1.6", "1.6 - 2.0", "2.0 - 2.4", "2.4 - 2.8", "2.8 - 3.2", "3.2 - 3.6", "3.6 - 4.0"], "values": [41432, 0, 981, 0, 0, 1221, 0, 502, 0, 1086]}',False,False
+capitalloss,0,0.0,4.0,0.1165583123258591,0.5605390667662566,0.0,0.0,0.0,integer,'[0, 1, 2, 3, 4]','{"keys": ["0.0 - 0.4", "0.4 - 0.8", "0.8 - 1.2", "1.2 - 1.6", "1.6 - 2.0", "2.0 - 2.4", "2.4 - 2.8", "2.8 - 3.2", "3.2 - 3.6", "3.6 - 4.0"], "values": [43082, 0, 360, 0, 0, 723, 0, 763, 0, 294]}',False,False
+hoursperweek,0,0.0,4.0,1.98805890938039,0.8752176175356808,2.0,2.0,2.0,integer,'[0, 1, 2, 3, 4]','{"keys": ["0.0 - 0.4", "0.4 - 0.8", "0.8 - 1.2", "1.2 - 1.6", "1.6 - 2.0", "2.0 - 2.4", "2.4 - 2.8", "2.8 - 3.2", "3.2 - 3.6", "3.6 - 4.0"], "values": [3602, 0, 5198, 0, 0, 26149, 0, 8684, 0, 1589]}',False,False
+native-country,0,,,,,,,,categorical,'["Cambodia", "Canada", "China", "Columbia", "Cuba", "Dominican-Republic", "Ecuador", "El-Salvador", "England", "France", "Germany", "Greece", "Guatemala", "Haiti", "Holand-Netherlands", "Honduras", "Hong", "Hungary", "India", "Iran", "Ireland", "Italy", "Jamaica", "Japan", "Laos", "Mexico", "Nicaragua", "Outlying-US(Guam-USVI-etc)", "Peru", "Philippines", "Poland", "Portugal", "Puerto-Rico", "Scotland", "South", "Taiwan", "Thailand", "Trinadad&Tobago", "United-States", "Vietnam", "Yugoslavia"]','{"keys": ["United-States", "Mexico", "Philippines", "Germany", "Puerto-Rico", "Canada", "El-Salvador", "India", "Cuba", "Others"], "values": [41292, 903, 283, 193, 175, 163, 147, 147, 133, 1786]}',False,False
+class,0,,,,,,,,binary,'["<=50K", ">50K"]','{"keys": ["<=50K", ">50K"], "values": [34014, 11208]}',True,False
